moved css files into assets folder.
